Moves the SQLQueryGenerator class to the database package, and removes some redundancy in portal removal
